cleaned up dtos
