+++ src/main/java/net/crusadergames/bugwars/dto/request/SignupRequest.java
@@ -1,4 +1,4 @@
-package net.crusadergames.bugwars.payload.request;
+package net.crusadergames.bugwars.dto.request;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
