--- src/main/java/net/crusadergames/bugwars/payload/response/MessageResponse.java
@@ -1,17 +1,0 @@
-package net.crusadergames.bugwars.payload.response;
-
-public class MessageResponse {
-  private String message;
-
-  public MessageResponse(String message) {
-    this.message = message;
-  }
-
-  public String getMessage() {
-    return message;
-  }
-
-  public void setMessage(String message) {
-    this.message = message;
-  }
-}
