+++ src/main/java/net/crusadergames/bugwars/service/AuthService.java
@@ -1,11 +1,11 @@
 package net.crusadergames.bugwars.service;
 
+import net.crusadergames.bugwars.dto.request.LoginRequest;
+import net.crusadergames.bugwars.dto.request.SignupRequest;
+import net.crusadergames.bugwars.dto.response.JwtResponse;
 import net.crusadergames.bugwars.model.auth.ERole;
 import net.crusadergames.bugwars.model.auth.Role;
 import net.crusadergames.bugwars.model.auth.User;
-import net.crusadergames.bugwars.payload.request.LoginRequest;
-import net.crusadergames.bugwars.payload.request.SignupRequest;
-import net.crusadergames.bugwars.payload.response.JwtResponse;
 import net.crusadergames.bugwars.repository.auth.RoleRepository;
 import net.crusadergames.bugwars.repository.auth.UserRepository;
 import net.crusadergames.bugwars.security.jwt.JwtUtils;
